refactor: modularize routing and ws handling
